% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contsimmap.R
\docType{package}
\name{contsimmap}
\alias{contsimmap-package}
\alias{contsimmap}
\title{contsimmap: Simulating conntinuous ancestral states from phylogenetic comparative data}
\description{
contsimmap more detailed description
}
\section{function category 1}{

functions...
}

