% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.checks.R
\name{.fix.trait.names}
\alias{.fix.trait.names}
\title{recycling explanation: Generally speaking, named entries of input will be matched up to specified trait/state names, then any entries with either no
or unmatched names are recycled in their given order to form an output of appropriate dimensions. When there are unmatched or unnamed entries, missing
entries are filled in with default values: 1 for variances and 0 for covariances (i.e., the identity matrix).}
\usage{
.fix.trait.names(k, nms)
}
\description{
In the case of matrices specifically, the function first attempts to "symmetrize" the input. For vector inputs, this creates a diagonal matrix. For
matrix inputs, rows/columns with missing names are labelled either with their corresponding column/row names (if they exist) or a numerical index.
Extra rows/columns of NAs are tacked on to the matrix so each row has a corresponding column and vice versa, the matrix is rearranged such that its row
and column names are identical, and all NA entries i,j are replaced with non-NA entries j,i. From here, the output is formed from the symmetrized input
as described above. Unmatched rows/columns of the input matrix are recycled into a block-diagonal matrix, and any unspecified covariances remaining are
set to 0.
}
